improves connection visibility in diagram compartments
Disables the default GMF compartment connection refresh manager in
DiagramEditParts as it hides connections that are only partially
visible. This might be necessary for compartments without an own
connection layer and needs to clip connections this way. However,
DiagramEditParts have their own child connection layer which is clipped
by draw2d and the old clipping technique is obsolete.