restricts point containment check to actual connection overlay outlines
restricts the point containment check (#containsPoint(x,y)) of
ChangeOverlayConnectionFigure and ConnectionChangeOutline to the actual
outlines. Otherwise the bounds have been used for this check and other
components might assume that the whole bounded area is covered by the
figure (which is not true as bounds can only be rectangular) - which
made selecting connection overlays impossible that were "covered" by the
bounds of other connection overlays.