fixes off screen indicator resolver behavior if indicator is not visible
the off screen indicator resolver did not consider the visibility of the
off screen indicator that has been clicked on. This resulted in
undesired scrolling when the user "accidently" clicked on an invisible
indicator (off screen indicators are invisible if they have been merged
or the figure they point to is visible). Fixed by disabling the
scrolling behavior if the indicator is invisible.