fixes default column comparators in review explorer and patchset history
the default comparators for the tree viewers in the review explorer and
patchset history view did not support tree viewers with multiple
columns. This caused string sorting based on the default label provider
(which uses #toString()). However, the desired default behavior should
be the same as if no sorting is applied to any column. Fixed by using
the comparator associated with the first column as default comparator.